requirements.txt
google-genai
flask
mysql-connector-python
python-dotenv
pandas